feat: Implement chase simulation mechanics
- Added chase simulation logic to allow players to pursue and capture prey based on their choices (Rock, Paper, Scissors).
- Introduced functions to determine if a player can eat another, find the nearest prey, and simulate chase ticks.
- Updated player structures and game state to accommodate float-based coordinates for movement.
- Modified server commands and state management to handle new chase mechanics.
- Enhanced client GUI to support auto-joining with specified choices and spawn locations.
- Updated tests to cover new chase simulation features and ensure correct behavior.

+++ src/client_gui_state.h
@@ -16,10 +16,10 @@ typedef struct
 	int used;
 	char name[MAX_NAME];
 
-	/* Last known R/P/S type and board position. */
+	/* Last known R/P/S type and board position (float-based coordinates). */
 	char choice;
-	int x;
-	int y;
+	float x;
+	float y;
 
 	/* Admission/round visibility flags from STATE snapshot. */
 	int alive;
